third_party/WebKit/LayoutTests/inspector/sources/debugger/navigator-view.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="../../../http/tests/inspector/workspace-test.js"></script>

<script>
function test()
{
    InspectorTest.createWorkspace(true);
    var sourcesNavigatorView = new WebInspector.SourcesNavigatorView();
    sourcesNavigatorView.setWorkspace(InspectorTest.testWorkspace);
    sourcesNavigatorView.show(WebInspector.inspectorView.element);
    var contentScriptsNavigatorView = new WebInspector.ContentScriptsNavigatorView();
    contentScriptsNavigatorView.setWorkspace(InspectorTest.testWorkspace);
    contentScriptsNavigatorView.show(WebInspector.inspectorView.element);

    var uiSourceCodes = [];
    function addUISourceCode(url, isContentScript)
    {
        var contentProvider = new WebInspector.StaticContentProvider(WebInspector.resourceTypes.Script, "");
        var uiSourceCode = InspectorTest.testNetworkWorkspaceBinding.addFileForURL(url, contentProvider, isContentScript);
        uiSourceCodes.push(uiSourceCode);
    }

    function revealUISourceCode(uiSourceCode)
    {
        sourcesNavigatorView.revealUISourceCode(uiSourceCode);
        contentScriptsNavigatorView.revealUISourceCode(uiSourceCode);
    }

    function dumpNavigator()
    {
        InspectorTest.dumpNavigatorView(sourcesNavigatorView, "sources", "  ");
        InspectorTest.dumpNavigatorView(contentScriptsNavigatorView, "contentScripts", "  ");
    }

    function expandDomains(tree)
    {
        var children = sourcesNavigatorView._scriptsTree.children;
        for (var i = 0; i < children.length; ++i)
          children[i].expand();
    }

    var rootURL = "http://localhost:8080/LayoutTests/inspector/debugger/";

    InspectorTest.addResult("Adding first resource:");
    addUISourceCode(rootURL + "foo/bar/script.js", false);
    dumpNavigator();
    InspectorTest.addResult("Expanding domains:");
    expandDomains();
    dumpNavigator();
    InspectorTest.addResult("Adding second resource:");
    addUISourceCode(rootURL + "foo/bar/script.js?a=2", false);
    dumpNavigator();
    InspectorTest.addResult("Adding other resources:");
    addUISourceCode(rootURL + "foo/bar/script.js?a=1", false);
    addUISourceCode(rootURL + "foo/baz/script.js", false);
    dumpNavigator();

    InspectorTest.addResult("Adding content scripts and some random resources:");
    addUISourceCode(rootURL + "foo/bar/contentScript2.js?a=1", true);
    addUISourceCode(rootURL + "foo/bar/contentScript.js?a=2", true);
    addUISourceCode(rootURL + "foo/bar/contentScript.js?a=1", true);
    addUISourceCode("http://example.com/", false);
    addUISourceCode("http://example.com/?a=b", false);
    addUISourceCode("http://example.com/the%2fdir/foo?bar=100&baz=a%20%2fb", false);
    addUISourceCode("http://example.com/the%2fdir/foo?bar=100%25&baz=a%20%2fb", false);
    addUISourceCode("http://example.com/path%20with%20spaces/white%20space.html", false);
    addUISourceCode("?a=b", false);
    addUISourceCode("very_looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong_url", false);
    dumpNavigator();

    InspectorTest.addResult("Revealing first resource:");
    revealUISourceCode(uiSourceCodes[0]);
    dumpNavigator();

    // Here we keep http://localhost:8080/LayoutTests/inspector/debugger2/ folder collapsed while adding resources into it.
    InspectorTest.addResult("Adding some resources to change the way debugger folder looks like, first:");
    var rootURL2 = "http://localhost:8080/LayoutTests/inspector/debugger2/";
    addUISourceCode(rootURL2 + "foo/bar/script.js", false);
    dumpNavigator();

    InspectorTest.addResult("Second:");
    addUISourceCode(rootURL2 + "foo/bar/script.js?a=2", false);
    dumpNavigator();
    InspectorTest.addResult("Others:");
    addUISourceCode(rootURL2 + "foo/bar/script.js?a=1", false);
    addUISourceCode(rootURL2 + "foo/baz/script.js", false);
    dumpNavigator();

    var rootURL3 = "http://localhost:8080/LayoutTests/inspector/debugger3/";
    addUISourceCode(rootURL3 + "hasOwnProperty/__proto__/constructor/foo.js", false);
    addUISourceCode(rootURL3 + "hasOwnProperty/__proto__/foo.js", false);
    addUISourceCode(rootURL3 + "hasOwnProperty/foo.js", false);
    dumpNavigator();

    InspectorTest.addResult("Revealing all resources:");
    for (var i = 0; i < uiSourceCodes.length; ++i)
        revealUISourceCode(uiSourceCodes[i]);
    dumpNavigator();

    InspectorTest.addResult("Removing all resources:");
    InspectorTest.testNetworkWorkspaceBinding.reset();
    dumpNavigator();

    InspectorTest.completeTest();
}
</script>

</head>
<body onload="runTest()">
<p>
Tests scripts panel file selectors.
</p>
</body>

</html>
